+++ src/main/java/net/geforcemods/securitycraft/SCTags.java
@@ -12,7 +12,6 @@ public class SCTags {
 		public static final TagKey<Block> REINFORCED_ACACIA_LOGS = tag("reinforced/acacia_logs");
 		public static final TagKey<Block> REINFORCED_BIRCH_LOGS = tag("reinforced/birch_logs");
 		public static final TagKey<Block> REINFORCED_BUTTONS = tag("reinforced/buttons");
-		public static final TagKey<Block> REINFORCED_WOOL_CARPETS = tag("reinforced/wool_carpets");
 		public static final TagKey<Block> REINFORCED_COBBLESTONE = tag("reinforced/cobblestone");
 		public static final TagKey<Block> REINFORCED_CRIMSON_STEMS = tag("reinforced/crimson_stems");
 		public static final TagKey<Block> REINFORCED_DARK_OAK_LOGS = tag("reinforced/dark_oak_logs");
@@ -21,6 +20,7 @@ public class SCTags {
 		public static final TagKey<Block> REINFORCED_ICE = tag("reinforced/ice");
 		public static final TagKey<Block> REINFORCED_JUNGLE_LOGS = tag("reinforced/jungle_logs");
 		public static final TagKey<Block> REINFORCED_LOGS = tag("reinforced/logs");
+		public static final TagKey<Block> REINFORCED_MANGROVE_LOGS = tag("reinforced/mangrove_logs");
 		public static final TagKey<Block> REINFORCED_NYLIUM = tag("reinforced/nylium");
 		public static final TagKey<Block> REINFORCED_OAK_LOGS = tag("reinforced/oak_logs");
 		public static final TagKey<Block> REINFORCED_PLANKS = tag("reinforced/planks");
@@ -35,6 +35,7 @@ public class SCTags {
 		public static final TagKey<Block> REINFORCED_WOODEN_SLABS = tag("reinforced/wooden_slabs");
 		public static final TagKey<Block> REINFORCED_WOODEN_STAIRS = tag("reinforced/wooden_stairs");
 		public static final TagKey<Block> REINFORCED_WOOL = tag("reinforced/wool");
+		public static final TagKey<Block> REINFORCED_WOOL_CARPETS = tag("reinforced/wool_carpets");
 		public static final TagKey<Block> SECRET_SIGNS = tag("secret_signs");
 		public static final TagKey<Block> SECRET_STANDING_SIGNS = tag("secret_standing_signs");
 		public static final TagKey<Block> SECRET_WALL_SIGNS = tag("secret_wall_signs");
@@ -50,7 +51,6 @@ public class SCTags {
 		public static final TagKey<Item> REINFORCED_ACACIA_LOGS = tag("reinforced/acacia_logs");
 		public static final TagKey<Item> REINFORCED_BIRCH_LOGS = tag("reinforced/birch_logs");
 		public static final TagKey<Item> REINFORCED_BUTTONS = tag("reinforced/buttons");
-		public static final TagKey<Item> REINFORCED_WOOL_CARPETS = tag("reinforced/wool_carpets");
 		public static final TagKey<Item> REINFORCED_COBBLESTONE = tag("reinforced/cobblestone");
 		public static final TagKey<Item> REINFORCED_CRIMSON_STEMS = tag("reinforced/crimson_stems");
 		public static final TagKey<Item> REINFORCED_DARK_OAK_LOGS = tag("reinforced/dark_oak_logs");
@@ -59,6 +59,7 @@ public class SCTags {
 		public static final TagKey<Item> REINFORCED_ICE = tag("reinforced/ice");
 		public static final TagKey<Item> REINFORCED_JUNGLE_LOGS = tag("reinforced/jungle_logs");
 		public static final TagKey<Item> REINFORCED_LOGS = tag("reinforced/logs");
+		public static final TagKey<Item> REINFORCED_MANGROVE_LOGS = tag("reinforced/mangrove_logs");
 		public static final TagKey<Item> REINFORCED_NYLIUM = tag("reinforced/nylium");
 		public static final TagKey<Item> REINFORCED_OAK_LOGS = tag("reinforced/oak_logs");
 		public static final TagKey<Item> REINFORCED_PLANKS = tag("reinforced/planks");
@@ -73,6 +74,7 @@ public class SCTags {
 		public static final TagKey<Item> REINFORCED_WOODEN_SLABS = tag("reinforced/wooden_slabs");
 		public static final TagKey<Item> REINFORCED_WOODEN_STAIRS = tag("reinforced/wooden_stairs");
 		public static final TagKey<Item> REINFORCED_WOOL = tag("reinforced/wool");
+		public static final TagKey<Item> REINFORCED_WOOL_CARPETS = tag("reinforced/wool_carpets");
 		public static final TagKey<Item> SECRET_SIGNS = tag("secret_signs");
 
 		private static TagKey<Item> tag(String name) {