+++ Changelog.md
@@ -1,5 +1,6 @@
 --------------------------Changelog for v1.9.3 of SecurityCraft--------------------------
 
+- New: Reinforced Blocks: Mangrove Planks, Mangrove Log, Stripped Mangrove Log, Stripped Mangrove Wood, Mangrove Wood, Mangrove Slab, Mangrove Stairs, Mangrove Button, Mangrove Pressure Plate, Mud, Packed Mud, Mud Bricks, Mud Brick Slab, Mud Brick Stairs, Mud Brick Wall, Ochre Froglight, Verdant Froglight, Pearlescent Froglight
 - New: Some of SecurityCraft's blocks, items, and entities can now trigger sculk sensors
 - Change: Reduced the Rail Mine's explosion size by 25% to make it comparable to other mines' explosions
 - Fix: Water doesn't render correctly while a Sonic Security System is recording within the player's view